todos api

+++ app/Http/Controllers/Api/ApiTodoIdController.php
@@ -74,10 +74,12 @@ class ApiTodoIdController extends Controller
         } // alles ok
         else {
             $todo = Todo::find($id);
-
-            $todo->update($request->validated());
-            $todo = new TodoResource($todo);
-
+            if($todo) {
+                $todo->update($request->validated());
+                $todo = new TodoResource($todo);
+            } else {
+                $todo = ['error' => 'not found'];
+            }
         }
 
         return response()->json($todo);
@@ -92,10 +94,16 @@ class ApiTodoIdController extends Controller
     public function destroy($id)
     {
         $todo = Todo::find($id);
-
-        $todo->delete();
-        $todo = new TodoResource($todo);
-
+        if($todo) {
+            $todo->delete();
+            $todo = new TodoResource($todo);
+        } else {
+            $todo = ['error' => 'not found'];
+        }
         return response()->json($todo);
+    }
+
+    public function error(){
+        return response()->json(['error' => 'route not found']);
     }
 }
